@@ -1,25 +1,37 @@
-allprojects {
-  buildscript {
-    repositories {
-      jcenter()
-    }
+buildscript {
+  repositories {
+    mavenCentral()
+    google()
   }
   dependencies {
-    repositories {
-      mavenCentral()
-    }
+    classpath libs.androidGradlePlugin
+    classpath libs.maven.publish.gradlePlugin
   }
+}
 
-  ext {
-    compileSdkVersion = 25
-    buildToolsVersion = '25.0.2'
-
-    supportLibVersion = '25.3.0'
-    butterKnifeVersion = '8.5.1'
-  }
-
+allprojects {
   group = GROUP
   version = VERSION_NAME
 
-  apply plugin: 'maven'
+  repositories {
+    mavenCentral()
+    google()
+  }
+}
+
+subprojects {
+  plugins.withId('com.android.base') {
+    android {
+      defaultConfig {
+        minSdk 14
+        compileSdk 34
+        targetSdkVersion 34
+      }
+
+      compileOptions {
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
+      }
+    }
+  }
 }
