@@ -1,8 +1,8 @@
 package com.mattprecious.telescope;
 
 import android.graphics.Bitmap;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import java.io.File;
 
 /**
